+++ src/tutorials/index.ts
@@ -1,19 +1,17 @@
-import {yellow, red} from 'chalk';
+import {red} from 'chalk';
 import findTutorials from './find-tutorials';
 import getPackageJson from '../packageJson/get';
 
-export default function tutorials(): string[] | boolean {
-  // console.log(yellow('This feature is not yet implemented'));
-
-  const pj: PackageJson = getPackageJson();
+export default function tutorials(dir: string): string[] {
+  const pj: PackageJson = getPackageJson(dir);
 
   if (!pj) {
-    console.log(red(`No package.json available`))
-    return false;
+    console.log(red(`No package.json available`));
+    return null;
   }
 
   return ([]
-    .concat(findTutorials(process.cwd(), pj.dependencies))
-    .concat(findTutorials(process.cwd(), pj.devDependencies))
+    .concat(findTutorials(dir, pj.dependencies))
+    .concat(findTutorials(dir, pj.devDependencies))
   );
 }