+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.6.0] - WIP
 - allow for programmatic use
+- working `> coderoad tutorials`
 
 ## [0.5.0] - 2016-05-02
 - bug fixes